+++ test/test_syscheck.js
@@ -312,6 +312,24 @@ describe('Syscheck', function() {
             });
         });
 
+        it('Filters: distinct', function(done) {
+            request(common.url)
+            .get("/syscheck/000?select=uid&distinct")
+            .auth(common.credentials.user, common.credentials.password)
+            .expect("Content-type",/json/)
+            .expect(200)
+            .end(function(err,res){
+                if (err) return done(err);
+
+                res.body.should.have.properties(['error', 'data']);
+
+                res.body.error.should.equal(0);
+                res.body.data.totalItems.should.equal(1);
+                res.body.data.items.should.be.instanceof(Array)
+                done();
+            });
+        });
+
     });  // GET/syscheck/:agent_id
 
     describe('GET/syscheck/:agent_id/last_scan', function() {